+++ tests/test_models/nn/test_patchts.py
@@ -9,20 +9,51 @@ from etna.transforms import StandardScalerTransform
 @pytest.mark.parametrize(
     "horizon",
     [
-        8
+        8,
+        13,
+        15
     ],
 )
-def test_rnn_model_run_weekly_overfit_without_scaler(ts_dataset_weekly_function_with_horizon, horizon):
+def test_patchts_model_run_weekly_overfit_with_scaler_small_patch(ts_dataset_weekly_function_with_horizon, horizon):
     ts_train, ts_test = ts_dataset_weekly_function_with_horizon(horizon)
-
+    std = StandardScalerTransform(in_column="target")
+    ts_train.fit_transform([std])
     encoder_length = 14
     decoder_length = 14
     model = PatchTSModel(
-        encoder_length=encoder_length, decoder_length=decoder_length, trainer_params=dict(max_epochs=10)
+        encoder_length=encoder_length, decoder_length=decoder_length, patch_len=1, trainer_params=dict(max_epochs=20)
     )
-    future = ts_train.make_future(horizon, tail_steps=encoder_length)
+    future = ts_train.make_future(horizon, transforms=[std], tail_steps=encoder_length)
     model.fit(ts_train)
     future = model.forecast(future, prediction_size=horizon)
+    future.inverse_transform([std])
 
     mae = MAE("macro")
-    assert mae(ts_test, future) < 0.06
+    assert mae(ts_test, future) < 0.9
+
+
+@pytest.mark.long_2
+@pytest.mark.parametrize(
+    "horizon",
+    [
+        8,
+        13,
+        15
+    ],
+)
+def test_patchts_model_run_weekly_overfit_with_scaler_medium_patch(ts_dataset_weekly_function_with_horizon, horizon):
+    ts_train, ts_test = ts_dataset_weekly_function_with_horizon(horizon)
+    std = StandardScalerTransform(in_column="target")
+    ts_train.fit_transform([std])
+    encoder_length = 14
+    decoder_length = 14
+    model = PatchTSModel(
+        encoder_length=encoder_length, decoder_length=decoder_length, trainer_params=dict(max_epochs=20)
+    )
+    future = ts_train.make_future(horizon, transforms=[std], tail_steps=encoder_length)
+    model.fit(ts_train)
+    future = model.forecast(future, prediction_size=horizon)
+    future.inverse_transform([std])
+
+    mae = MAE("macro")
+    assert mae(ts_test, future) < 1.3